requirements.txt
passlib[bcrypt]
python-jose[cryptography]
fastapi>=0.95
uvicorn[standard]>=0.22
SQLAlchemy>=1.4
pydantic>=1.10
pydantic_settings
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
